+++ src/ansible-webui/aw/config/hardcoded.py
@@ -10,3 +10,10 @@ FILE_TIME_FORMAT = '%Y-%m-%d_%H-%M-%S'
 KEY_TIME_FORMAT = '%Y-%m-%d-%H-%M-%S'
 MIN_SECRET_LEN = 30
 JOB_EXECUTION_LIMIT = 20
+GRP_MANAGER = {
+    'job': 'AW Job Managers',
+    'permission': 'AW Permission Managers',
+    'repository': 'AW Repository Managers',
+    'credentials': 'AW Credentials Managers',
+    'system': 'AW System Managers',
+}